Added the repeted classes like material, bearing discrete and nuts_and_bolts_dimensions in a new commonComponents.py file
